Ingest user-provided metadata line-by-line
This replaces the previous functionality which required a user to upload
an entire, likely sparse, DataFrame to edit metadata. Previously, a user
had to create a new column in the TSV for every unique field they wanted
to edit.

Now, annotations can be done in the following format:
(no header)
{strain name}   {column name}   {intended value}