Add a `slack` option when fetching data
Add a new option, `--slack`, for the `fetch-data` script. I realized
that users running this script locally probably don't want a `@channel`
notification sent, creating noise. The purpose of the Slack notification
is for the cron job/automated user to notify us of changes in the
dataset.

Update the GitHub workflow to use the new `--slack` option. In the
future, we may want to rewrite the `fetch-data` script as separate
scripts, or move parts of it (like the Slack notification) to the `run`
portion of a GitHub actions job/step.