Refactor fetch-data script into several smaller, single-purpose programs
These are more composable and increase the separation of concerns,
making it easier to test individual parts in dev and change the
behaviour of once piece without affecting the rest.  For example, Slack
notifications now care about record count, not SHA-256 hashes.

Introduces a portable sha256sum program so the implementation in GNU
coreutils isn't necessary.