Notify on additional info change
Check for differences between the local `additional_info.tsv` and the
one on S3. Filter out empty rows from `additional_host_info` in both
TSVs to reduce noise when adding new sequences without relevant
additional data.

Note that the current process relies on the fact that column 3 is
`additional_host_info`. Make a note in the `transform` script to
preserve the current ordering of columns. Ideally we would verify that
the first two columns are our IDs (`gisaid_epi_isl` and `strain`) and
filter out all the remaining columns with empty values, but I'm new to
`awk` and don't know how to achieve this now.