Revert "transform: Sort output data by `gisaid_epi_isl`"
This reverts commit fc22db67f141c3520bf010ae2a9949f2a89312b8.

Trevor finds "it easier to look at the group of Swiss sequences or WA
sequences or whatever. Strain grouping helps with this."

This will cause rebuilds of staging when just strain names change, but
that's ok (and maybe preferable to get the new names anyway).

Note that while the reverted commit implies strain isn't unique, it is
by this point due to the drop_duplicate() in the code immediately
preceding.