Ensure region_exposure is set for all strains which have country_exposure set
This doesnt technically need to be present if the region_exposure happens to match the region, but problems arise if this is not the case.
To make this clear, region_exposure should be explicitly set for all strains which have country_exposure set.
This has been done for all data in this commit.
Most were a result of simple spelling errors (e.g. "regions_exposure")

A script ("./bin/check-annotations") has been added to identify such strains