Flag future dates
Sometimes we received data from GISAID with collection dates in the
future. This causes a big headache for the Nextstrain nCoV build
maintainers when left uncaught until the end of a multi hour long build.

We can exclude future dates from builds downstream with `augur filter
--max-date`, therefore, there is no need to add sequences with future
dates to an _exclude.txt_. We do, however, still want to warn nCoV build
maintainers about these bad metadata, so we should still maintain a list
of flagged metadata to send to Slack.

The Slack message contains the GISAID EPI ISL ID and collection date for
every strain that has a collection date from the future (e.g.  a date >
now). Emma specifically requested that this Slack message be formatted
differently from the others so that it stands out, so send an additional
Slack message with an emoji   and some text alerting the user to the new
exclusions. I could have edited the `notify-slack` script to accept an
optional text argument when `--upload` is used, but it seemed easier to
just send two Slack messages instead.