Fetch all GenBank sequences + metadata from NCBI Virus
This adds a new file, genbank.ndjson.gz, which currently contains about
~3,600 sequences.  The previous genbank.fasta.gz file will no longer be
updated; it was a small subset of just 62 sequences which we'd manually
linked between GISAID and GenBank.  58 of those sequences are in the new
file; the remainder were removed from GenBank in the interim.

Note that the GenBank data is still not ingested, just fetched and
tracked.  The next steps are to incorporate it into the "transform" step
such that it ends up mixed in with the GISAID data in metadata.tsv and
sequences.fasta by:

  1. Aligning the fields and values between GISAID and GenBank

  2. Matching sequences in both repositories, likely by strain name
     heuristics, and then deduplicating matches.  Emma has already started
     scoping out the feasibility of this.