fetch-from-genbank: Add additional isolate and region fields
Discovered by Jover! \o/

For all existing records, isolate appears to be the same isolate name
that's embedded in the larger title.  We could possibly drop the title
field entirely, but I'm not sure if the modicum of extra information
there is ever useful to humans looking at the data or not, so I'm
leaving it.  I moved title to the end to keep the best fields
front-loaded (sequence is the very best field, but it's last because its
values are so long).