transform: Reduce mangling of author and lab names
Preserves accented characters and other non-ASCII marks instead of
folding and stripping everything to ASCII.

The construction of "et al" forms for author had to change because they
relied on ASCII as well as incorrect assumptions about the structure of
names.  Handling is much simpler now and names are less mangled or
mistakenly lumped together, although there are still issues, as noted.