launch-build: Use a larger node to speed up tree building
Targets the c5.24xlarge instance size with 96 cores, which costs
$4.08/hr on demand.  Overrides the hardcoded thread count for the tree
rule in the Snakemake profile used.

This lets the trees run in parallel even though 16 × 7 > 96 because some
of the tree jobs are very short and others are very long.  Total runtime
in a test build was about 2.5 hours.¹

Because of fixes to pathological memory usage made a few weeks ago, we
no longer need the high memory z1d nodes (although they still have the
fastest CPUs).

I may experiment with this some more and tweak node sizing again in the
near future.

¹ <https://blab.github.io/snakemake-stats/?stats=https://tsibley.net/paste/2020-05-28YXMQRdFY-ncov-stats-cpus=96-threads:tree=16-instance=c5.24xlarge.json.json>