Limit what kinds of pushes trigger the ingest workflow
Pushes to master now only trigger the workflow if our manual annotations
are modified.  Code changes won't trigger it anymore.  Now that we're
aiming for a less rapid pace of updates—scheduled builds are only every
6 hours—this makes more sense than triggering builds off that schedule.
Ingest on master can still be triggered manually with `./bin/trigger
ingest`.

Pushes to branches other than master still always run, with the idea
that this is useful for testing during development.

The easiest way to express this without writing a bit of code to parse
the GitHub event payload and `git log --stat` was to split the workflow
definition into two and use the built-in push filtering affordances.