transform-genbank: update parsing of geographic columns
NCBI does not verify the free text sub-regions for locations. They
recommend annotating sub-regions from largest to smallest, but this is
not strictly enforced.

From digging into the avaiable sequences, it seems submissions from
outside the US follows this format. Submissions from within the US
varies in format, so it requires some special checks to verify that
the state name is correctly parsed into the `division` field rather
than the `location` field.