utils: Don't evaluate empty column's dtype
In `standardize_dataframe()`, we create empty columns and fill them with
`pd.NA` which results in a numpy.dtype that is not comparable with
panda's `string` dtype. This resulted in a TypeError when evaluating if
the empty column's dtype `== 'string'`.

Because we only use `pd.NA` when creating empty columns, check if the
column is not entirely empty before comparing its dtype to panda's
`string`.