ingest-gisaid: Always use unix line endings
PR #90 recently introduced a major refactoring and optimization of
`transform-gisaid` that has resulted in producing Windows (CRLF) line
endings in the metadata and additional info output TSVs. This was
somehow not detected in local testing where output files were shown to
be identical between the original and new transform scripts. As a
temporary fix to keep the ingest operational, always use the
`--output-unix-newline` option when invoking `transform-gisaid` in
`ingest-gisaid`.