clean-tsv: Stop reading in TSV twice
Read in the target TSV only once, optionally enforcing a certain number
of columns with the --n-cols option. Achieve this by always reading in
the TSV with header=None to enforce an expected number of columns when
given --n-cols. Then, after reading in the data, if a header should be
used (i.e. --no-header is not specified), replace the column names with
the first row.