fix: keep "n/a" values in metadata
A followup of #100

By default, Pandas parses `NA`, `n/a` and similar values as `NaN` when reading a TSV file and then writes an empty cell when saving.

This PR disables this behavior, so that the values above are interpreted as strings and are preserved in the output as is.