bin/transform-gisaid address #103 strain name bug;
this is an attempt to fix a bug
where the strain names output
by bin/transform were not consistent
across metadata and fasta output files.
Since this was only an issue when
--sorted-fasta was not passed, this
addresses the issue by adding logic
to use strain names from the transformed
metadata list (to match the metadata output)
when writing out an unsorted fasta.
Performance hasn't been tested.