Lift-and-shift ingest pipelines to AWS Batch via `nextstrain build`
It's been looming on the horizon for a while, but the impetus to move
off GitHub Actions for the bulk of this processing came to a head with
recent out-of-disk-space issues as sequence counts spiked.  We started
with GitHUb Actions last winter because it was expedient and simple to
do so, but now it seems more headache than it's worth.  Collateral
benefits include being able to provide the ingest with more CPUs and
memory to make them quicker.

Future improvements could/should include:

  - Defining the workflow via Snakemake, like typical Nextstrain builds.
    Richard and Eli have started work on this.

  - Short-term: Automatically updating and publishing the custom
    container image when deps change in git (using a new GitHub Actions
    workflow).

  - Longer-term: Replacing the custom container image with top-notch
    Nextstrain CLI support for satisfying arbitrary build dependencies
    (i.e. those in addition to Nextstrain core).