Explicitly depend on envdir
It's required for launching our ingests via `nextstrain build` and
*should* be in the container image, but this morning's run found it
missing due to nextstrain/base:branch-python-base not being rebuilt on
top of the most recent nextstrain/base:latest which included envdir.

I'll fix up the image situation too, but this change should provide a
quick fix and it shouldn't hurt to be explicit.