rebuild-staging: Fix Snakemake invocation for secondary test/trial build
Using --configfile seems to (?) override the profile's own config
instead of overlaying/merging with it, which resulted in this error:

    MissingRuleException: No rule to produce "upload" (if you use input
    functions make sure that they don't raise unexpected exceptions).

I was able to reproduce locally using the exact same Snakemake
invocation via the Nextstrain CLI and then inspect what was happening
both via --dry-run and the "dump_config" rule.  It did not seem to
matter if --configfile came before or after --profile.

Switch to using --config instead, which was actually my original
inclination as the (now removed) comment attests to.  I'd misunderstood
the behaviour of --config in my previous tests, which misled me to
believe it wouldn't work for our purposes.  I'd been testing by passing
--config multiple times, once per variable to set, which resulted in
just one of the variables set.  I chalked this up to differences in the
default values for those variables in the profile config files: one is
None, the other has a string value.  It looked to me like --config was
overwriting the None-valued variable but not the string-valued variable.
Weird behaviour, but not implausible.  What was really happening is that
the second --config was negating the first --config!  The --config
option is supposed to be given just once with a list of variables to
set.  Invoked this way, local testing shows that both variables override
the defaults provided by the profile.