flag-metadata: Flag collection dates prior to March 2020
While there are many valid sequences with such dates already in this
dataset, vanishingly few _new_ sequences are likely to appear with such
dates.  They're more likely to be typos for early 2021 dates, so the
build shepherds want notice of them.

This will change the steady state of flagged_metadata.txt from empty to
several thousand lines, but that should be ok, I think.