Tone down the Slack notification on new/changed data
While new data was high urgency at the beginning of this pandemic and
cause for immediate rebuilds, our build shepherds have long-settled into
a regular cadence where new data is processed on our own schedule
instead of as interrupts.  Mentioning @channel also masks "real"
mentions of individuals in the channel, which makes it harder to get the
attention of people tagged into conversations.