fetch-from-genbank: Use curl instead of the requests library
Between 31 May and 7 June, the fetch started failing with a cascade of
errors (in order):

    http.client.IncompleteRead: IncompleteRead(0 bytes read, 512 more expected)
    urllib3.exceptions.ProtocolError: ('Connection broken: IncompleteRead…')
    requests.exceptions.ChunkedEncodingError

I dug into this a bit yesterday and today and isolated it to something
in the network transfer itself (as opposed to the CSV parsing or
something, but see also below), but didn't dig further into what.  When
the process stalls, there's still transfer happening on the network, but
nothing is making it through the response body iterator.  It's unclear
why without more instrumenting, but it appears to be a bug in the
requests Python library (or underlying urllib3 or httplib2), as I can
complete the fetch successfully with curl.  Instead of going down this
particular rabbit hole, just switch over to curl.  There's no good
reason we have to use the requests library.

The request URL is still generated in Python in order to minimize
changes and keep the clarity of the existing parameters data structure
(as opposed to using curl's --data-urlencode option or hardcoding the
fully-encoded URL).

Note also that with curl I sometimes observed HTTP/2 framing errors
reported via its exit status, although the download always seemed to
complete successfully.  Nevertheless, disable HTTP/2 and use good old
HTTP/1.1 instead.

Once the network issue was resolved, the CSV max field size also needed
to be bumped up to accommodate a 114M base pair long sequence in the
results.  This is clearly not SARS-CoV-2 and the record was retracted
from the INSDC databases by the ENA, but it still exists in NCBI Virus.
(I'll write to the NCBI folks about this separately.)