Update rebuild script for GISAID and GenBank builds
The ncov/nextstrain_profiles/nextstrain profile is soon to be replaced
with two profiles (nextstrain-open and nextstrain-gisaid) representing
different data sources. (We have already created separate ingest
actions for these data sources.) This commit allows the rebuild
workflow to run the nCoV pipeline for both profiles.

This changes the way ./bin/rebuild is called, now requiring a single
argument specifying the data source to use. This removes the previous
ability to run the command with 4 arguments; a search of this repo
did not identify any code using this or any documentation utilizing it.

We may wish to replace the "rebuild" workflow with two workflows:
"rebuild-genbank" and "rebuild-gisaid" so that they can be triggered
independently when their respective data updates.