Automatically retry fetches from GISAID and GenBank
Try to recover from transient network/connection failures we sometimes
see.  Most of these are protocol errors or connections unexpected closed
by the remote peer.  We can't use curl's --retry option because it
doesn't apply to the kinds of errors we see and its --retry-all-errors
option would produce duplicate/corrupted data since we process it as a
stream and write to disk.

This retry logic is rather crude but should be fine for now.  At some
point when this pipeline uses a workflow engine like Snakemake,
presumably we can make use of builtin retry logic.