GenbankProblematicFilter: move strain side effect
In cases where we cannot parse the standard strain name from the GenBank
record, we default to the GenBank accession as the strain name. This
allows us to include GenBank sequences even if they do not have a
standardized strain name.

Move this change from `GenbankProblematicFilter` to
`StandardizeGenbankStrainNames` so that all strain name transformations
are contained within the same transform step.