transform-biosample: update submitting_lab for EBI
BioSample records pulled from EBI/ENA have owners as "EBI" or
"European Bioinformatics Institute" but we want to acknowledge the
original submitting lab that had submitted the sequences to EBI/ENA.

For records with "EBI" as the owner, the original submitter is stored
in the `INSDC center name` attribute.

For records with "European Bioinformatics Institute" as the owner, the
original submitter is stored in the `sampleIds` field within `db`
of the sample id that is labeled as "Sample name".