Trigger preprocessing pipelines after fetch-and-ingest
Our intention is to have the preprocessing (alignment) pipeline
run following fetch-and-ingest, which currently happens daily
from the master branch.

The trigger occurs from the bash scripts themselves to cover the
situation when the AWS job exceeds the GitHub action's run time limit
(6 hours), and thus we cannot rely on a subsequent GH Action step
to make the repository dispatch.

We currently have no automated method to trigger a preprocessing
action to use the assets produced by a fetch-and-ingest run
from a branch, however this can be accomplished by triggering
AWS jobs from your own machine.