Add `--nextclade_full_run` flag
This PR adds `--nextclade_full_run` flag for `ingest-gisaid` and `ingest-genbank` scripts. This flag is needed to trigger full run of Nextclade on all available sequences.

In current implementation, clade results (and other metrics) are cached in `nextclade.tsv` file. Upon rerun, the sequences in `sequences.fasta` and  `nextclade.tsv` files are compared and only sequences that are missing in `nextclade.tsv` undergo analysis with Nextclade. The results are then appended to `nextclade.tsv`. Then `nextclade.tsv` gets merged with `metadata.tsv` and that's how clades and some of the other Nextclade metrics appear in the medatata.

With the new flag `--nextclade_full_run` is provided, `nextclade.tsv` does not get downloaded. So the comparison always gives a list of all sequences and Nextclade processes them all again, and metadata table gets refreshed clades and metrics for all rows.

This refresh is necessary every time when new clades are defined or when upgrading to the new version of Nextclade with breaking changes.

Note: this might take very long time to run and requires substantial computational resources and fast I/O.

Previously we used to perform these runs manually on external resources, but now we are thinking about automating this.

The trigger should probably still be manual, because not every Nextclade release needs the full run.

Depending on the manual data curation flow, these runs can be performed on master or on a separate branch. If on separate branch, the resulting `metadata.tsv` and `nextclade.tsv` will end up in the `branch/` subdirectory and should be copied manually to the location when the next ingest from master can find them.

Another challenge arises when we want to perform this full run before the required Nextclade version is released. Nextclade currently does not publishes pre-releases of CLI, but that can be worked out. Although automated C++ builds from master branch might be too costly in terms of CI credits.

Let's experiment a little and see how well it goes.