[no ci] Add new GitHub Actions for full Nextclade runs
These are workflows from https://github.com/nextstrain/ncov-ingest/pull/215. I want to experiment with different options for full Nextclade runs, as described there. The `.github/workflows/*` files need to be on master in oreder for GitHub actions to be triggerable.

Note that the branch is currently hardcoded. We need the code from the branch to be able to trigger full runs.

```
    - uses: actions/checkout@v2
      with:
        ref: 'nextclade-full-run-2021-09-30'
```