feat: cleanup files when done
We are unlikely to require the uploaded working directory, so removing large files in the and and avoiding the zipping and uploading might save a few bucks, given the price category of resources. Also it's easier to reason about the running time of Nextclade when only Nextclade is running: compression of working dir takes more time than the run itself.