fix: empty lines after tsv concatenation
Looks like out concatenation script produces empty lines in the end of the file randomly.

Let's use use `csvstack` to concatenate tsv batches. `csvstack` comes from `csvkit`, which is a well-known package to handle CSV files and should be better that a hand-rolled solution. Most importantly, no empty lines. And is seems to be faster.