Snakemake pipeline
This commit introduces a Snakemake workflow to replace the scripts
`bin/ingest-genbank` and `bin/ingest-gisaid`. The aim is for the
pipeline behavior to be unchanged. This move sets us up to make
future changes such as improving slack notifications, implementing
cached alignments or combining this repo with nextstrain/ncov.

The pipeline was implemented with an aim of minimal (if any) changes to
the files in `./bin/`; these can come in subsequent PRs, as desired.
Similarly, the existing GitHub actions should behave as before.