feat: allow for custom Nextclade binary and dataset
This adds 2 checks in the beginning of "Nextclade full run" script:

1. Whether `nextclade` binary present in the root of the project.

   By default the latest release of Nextclade CLI is downloaded. But if the `nextclade` file is detected, then the download is skipped and this file is used instead.

2. `tmp/dataset/` directory containing `tree.json` and `reference.fasta` files.

   By default the latest dataset is downloaded. Hoverver, if a dataset is detected in the `tmp/dataset/`, the download is skipped and this dataset is used instead.

This allows to customize manual Nextclade full runs, when a combination of a specific Nextclade version (perhaps unreleased) and of a specific dataset is needed.

This customization is intended for cases when putting Nextclade binary and dataset into a local working directory of ncov-ingest and running one of the `./run-nextclade-full*` scripts manually. It would be nice to somehow simplify and automate this process, but for now I don't have good ideas for this. In particular, because the binary and datasets are different depending on a particular use-case for these custom runs.