Run ./bin/clean by default
This commit is in response to PR review comment
https://github.com/nextstrain/ncov-ingest/pull/231#discussion_r760553194

Note that this behavior is a change from the previous bash
implementation where an error would cause the script to exit before the
cleaning step.

File cleanup etc should be reviewed in a future PR - currently we use
snakemake's `temp()` and the `./bin/clean script (which is controlled
by this new config parameter; I tried to condition this on the
command line argument `--notemp` but could not find where, if anywhere,
snakemake exposes this).