chore: cleanup docker container image
This:

 - Removes duplicated nextclade downloads. Nobody knows wha version they are and when they are updated. Latest version of nextclade is already downloaded in the `./bin/run-nextclade` and `./bin/run-nextclade-fill` scripts. We want it to always be cutting edge. Nextclade releases are frequent. We may consider bringing it back into the image later if we setup automated rebuild of the image on Nextclade release. However I also often play with different custom versions of Nextclade in full runs and having other versions duplicated versions may cause mistakes. Mixed feelings here.

 - Adds various tools we might want to play with on branches (related to https://github.com/nextstrain/ncov-ingest/pull/247 https://github.com/nextstrain/ncov-ingest/pull/256):
    - aria2 - for downloading on multiple connections
    - lbzip2  - for parallel unbzipping
    - pigz - for parallel gz/ungz
    - pixz - for parallel xz/unxz

 - Changes base image to as suggested in https://github.com/nextstrain/ncov-ingest/pull/253. We don't use much out of it, so I am not super convinced it's needed, but not a big deal.

Current inconvenience is that for any changes in deps on branches (e.g. if I want to play with aria2c and lbzip2) we need to change the container on master and/or publish an image and change the tag in the GitHub Action on master. Which both disrupt production. We might consider installing deps dynamically on AWS batch instead. I tired and it failed because currently the debian repo cache in the image is too old. Or maybe there's a better way to play with new tools right now? In any case, mixing prod and dev as it is now seems weird and unproductive, we might think about a dedicated, simpler dev flow.