Build and push nextstrain/ncov-ingest:branch-* images
For use in dev and testing.  Parallels the pattern used by the
nextstrain/base images.

I was going to use the GitHub Action docker/metadata-action@v3 to
generate tags, but that started becoming more complicated than a small
bit of shell to set an env var.

With per-branch images, we could now also adjust the per-branch ingest
workflows use these images for the AWS Batch jobs (if a branch image
exists).