perf: let's try parallel xz and gz
This may or may not give a little speedup.

pixz has been a parallel alternative to xz for some time before xz finally implemented -T flag for compression. And it has parallel decompression. We might compare how the 2 perform.

pigz is a parallel gz from the author of the algorithm himself. Definitely faster, but we only use gz on smaller files, so not necessarily noticeable.

Both pixz and pigz are compatible with their respective "default" tool, in both directions.