rule notify_genbank: remove metadata diff
The `csv-diff` within `notify-on-metadata-change` began running into
a MemoryError on December 26, 2021. This was patched during the holidays
by increasing the memory for Genbank ingest runs (bd2c2ca).

Further investigation showed that we have not been receiving the
metadata diff output files since we switched to Snakemake on
December 1, 2021 (74de8b1). The `upload` rule has been running before
`notify_genbank` resulting in diffing of the same file that has been
uploaded to S3. This is caused by `upload` not having any dependencies
on `notify_genbank` and `notify_genbank` having an additional dependency
on `check_locations`. This prompted an internal discussion that showed
no one has been using the outputs of this metadata diff.