chore: document gha delays
See the modified comment for explanation.

I added a note about delays in GitHub Actions, as well as a reminder to modify the jobs for the 2 databases consistently.

I just want this info to be there, so that everyone who modifies schedule or waits for scheduled jobs is aware of these peculiarities.