Add MaskBadCollectionDate to gisaid and genbank transforms
Masks the collection date with `XXXX-XX-XX` if the collection date is
the same day as or after the submission date.

This was suggested by @trvrb when @corneliusroemer noticed that a subset
of GISAID data had collection dates identical to the submitting date.
It is most likely impossible to sample, PCR, sequence, and submit on
same day. However, if we do determine this is possible, the masking
can be overridden for specific cases via the manual annotations.