Notify Slack when GitHub Actions fail
When GitHub Actions fail, they automatically send a notification to the
GitHub user who triggered the action or last editor of the workflow
cron schedule¹. This commit updates the GitHub Action workflows to
notify the whole team via Slack notification if there is a failure in
the workflow.

This is separate from the Snakemake `onerror` handler because it catches
failures that are not within the Snakemake pipeline.

¹ https://docs.github.com/en/actions/monitoring-and-troubleshooting-workflows/notifications-for-workflow-runs