Add notify-on-job-fail
Sends a Slack notification when there are errors in the ingest pipeline.
The Slack notification includes the AWS Batch job id in Slack message
if it is available via the environment variable `AWS_BATCH_JOB_ID` or
the GitHub Action run id if it is available via the environment variable
`GITHUB_RUN_ID`.

Unable to provide details on the failure if neither environment variable
are available, but still sends a general failure message.