This removes the automated on push running of ingest
Currently, any branch updates result in running the entire GISAID and GenBank ingest workflows via GitHub Actions. In current build schedule and curation routine this is no longer necessary.

This commit removes the "on push" trigger, but leaves the branch-specific workflows intact via the "workflow_dispatch" option. The idea would be to be able to run trial ingests when desired. Rather than having it happen with every push.