notify-on-record-change: notify on all changes
We've been having stale fetches from the GISAID endpoint that lack new
records. It would be helpful to have specific notifications about the
different types of record changes, not just new records. This way we get
early notifications about the status of our fetches instead of having
to look for messages buried in the pipeline logs.