fetch_sequences: decouple AWS S3 fetch from database fetch
Only include rules for fetching sequences/metadata from AWS S3 if the
config `s3_dst` and `s3_src` params are provided.

Since the fetch from the database and the fetch from S3 produce the same
output, we define `ruleorder` to handle the `AmbiguousRuleException`.
We give precedence to fetching directly from the databases if
`fetch_from_database` is set to True.