Snakefile: make upload and trigger optional
Only include outputs/rules for upload if `s3_dst` is provided in config.
Only include outputs/rules for trigger if uploading files since the
trigger rules depend on the outputs from upload.

The default outputs of the main Snakemake pipeline are now just the
metadata.tsv and sequences.fasta files.