Split Snakefile into workflow modules
Split the Snakefile into 6 different modules to clean up the main
Snakefile and make it easier to understand the purpose of the rules.
Added documentation to the top of each module to explain the purpose
of each module and to document the expected inputs and outputs.

This is done in anticipation that we will make a subset of modules of
the pipeline optional so that external users can use the main workflow.