nextclade: decouple nextclade outputs from S3 uploads
Previously, the `aligned.fasta` file only got generated when we include
uploads to S3. This commit includes this file in the input of the
`generate_metadata` rule so that it is pulled through the Snakemake DAG
to be produced as an output whenever there are new sequences run through
NextClade.