transform-genbank: Add CLI options for COG-UK data
Add CLI options for the COG-UK sample accessions and metadata files
instead of hard-coding the URLs within the transform pipeline.
The default values for the options are the original hard-coded URLs so
the behavior is still the same when the new options are not used.

This opens the path for us to fetch the COG-UK files prior to running
the transform script so that the transform script does not run into
network errors.