feat: scrap checkpoint
Removed checkpoint because it only slows down the pipeline without
any of the past benefits. It was initially added because Nextclade v1
used to error on an empty FASTA input and we needed the checkpoint to
check if we should generate a new mutation summary¹. Now, running
Nextclade v2 on a no-op file takes ~1 minute and we no longer use the
mutation summary.

¹ https://github.com/nextstrain/ncov-ingest/pull/237#issuecomment-993674154