feat: remove threads for run_nextclade rule
We usually run ingest with 16 CPUs, assigning 64 threads means that this
rule cannot run in parallel to other rules which slows down ingest.

Since we use incremental alignments, the alignment part is fast, so we
can just let Nextclade use all cores available to get it done with -
this may be slightly subideal from a performance point of view but
fine-tuning threads for different parallel rules is a nightmare and
not worth it.

For the foreseeable future no problem will arise from parallelizing
this rule.