Trigger ncov rebuilds for new data only
We do not want to waste time and resources to run rebuilds when the
ncov-ingest outputs have not changed. This commit updates the pipeline
to check the outputs from the uploads of the metadata TSV and sequences
FASTA files used by downstream Nextstrain ncov builds [1,2] to verify
they have been updated on S3. If the output includes the message
"files are identical" then we skip the trigger for rebuilds.

[1] https://github.com/nextstrain/ncov/blob/89c1d18751d005c7d3869e93de8c2baa0407ba8a/nextstrain_profiles/nextstrain-gisaid/builds.yaml#LL32-LL33
[2] https://github.com/nextstrain/ncov/blob/89c1d18751d005c7d3869e93de8c2baa0407ba8a/nextstrain_profiles/nextstrain-open/builds.yaml#L32-L33