nextclade: switch from `cp` to `ln` for full run
In the latest "full" Nextclade run for GISAID, the pipeline failed with
a "No space left on device" error for rule combine_alignments. This was
unexpected considering the AWS EC2 instances have 2TB of disk space.¹

The easiest change in the Nextclade workflow rules is to switch from
`cp` to `ln` for full runs so that we don't have two copies of the full
GISAID sequences.fasta file, which will only continue to grow.

¹ https://bedfordlab.slack.com/archives/C01LCTT7JNN/p1647477064627379?thread_ts=1647353315.650509&cid=C01LCTT7JNN