join-metadata-and-clades: use `low_memory` for clades
In memory profiling with `filprofiler`, the major consumer of memory was
surprisingly the reading of the nextclade_21L file. Setting
`low_memory=True` lowers the memory usage and we do not have to worry
about mixed types since we are using the `dtype` parameter.

Sets `dtype="object"` for the main nextclade.tsv to avoid the mixed
type warnings. Future improvement can include specifying dtypes for
individual columns.