transforms/AbbreviateAuthors: remove double quotes
For some reason, certain author values from GISAID are double quoted
and this leaves an unclosed quote when we truncate the authors to
`"...et al`. This unclosed quote can lead to errors in table readers
such as Excel as reported in #395.

This commit strips all `"` from the truncated authors string as an
easy and quick fix for the issue.