Add more clade columns to metadata
We have updated Nextclade to separate out clade names (`23A`, `Omicron`, `23A (Omicron)`), but we have never followed up by doing the same in ingest, to give us more columns both to work with, and some overlap before we potentially change what `Nextstrain_clade` is. 

This also adds a column `display_clade` which currently is the legacy clade (`23A (Omicron)`) but could be changed to reflect whatever we want to display, and be guaranteed to be 'unstable' (don't base any workflows on this).