Dockerfile: Remove ~/.cache dirs created by Pip and Pipenv
This directory was getting created as root during image build which then
made it unwriteable by non-root users at run time.  This wasn't an issue
for a while, but Snakemake, as of newer versions¹, now uses the same
user cache dir for its own subdir (~/.cache/snakemake, i.e.
/nextstrain/.cache/snakemake).  Snakemake fails to create its own subdir
if the user cache dir is owned by root.

Like most in-image caching, these are not useful and only serve to
increase image size.

It's easiest to remove the directories explicitly rather than disable
their creation (possible for Pip with --no-cache-dir or
PIP_NO_CACHE_DIR=1, not possible for Pipenv).

An alternative would be to set HOME differently during build, but that
seems like a potential source of other unforeseen issues.  Maybe in the
future when we have a more general "derivative image" recipe.

¹ <https://github.com/nextstrain/docker-base/issues/135>
  <https://github.com/nextstrain/docker-base/pull/136>